pyproject.toml
[tool.poetry]
name = "orb-task-02"
version = "0.1.0"
description = "ORB CASE"
authors = ["Emre MENTEŞE <emrementese@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"
django-rest-knox = "^5.0.2"
model-bakery = "^1.20.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
